+++ i4j-fits-utils/src/main/java/org/indilib/i4j/fits/debayer/SmoothDebayerAlgorithm.java
@@ -71,13 +71,13 @@ class SmoothDebayerAlgorithm implements DebayerAlgorithm {
                     G3 = ip.getPixel(x + 1, y + 1);
                     G4 = ip.getPixel(x + 1, y - 1);
 
-                    g.putPixel(x, y, (int) G1);
+                    g.putPixel(x, y,  G1);
                     if (y == 0)
-                        g.putPixel(x + 1, y, (int) ((G1 + G2 + G3) / 3));
+                        g.putPixel(x + 1, y,  ((G1 + G2 + G3) / 3));
                     else
-                        g.putPixel(x + 1, y, (int) ((G1 + G2 + G3 + G4) / 4));
+                        g.putPixel(x + 1, y,  ((G1 + G2 + G3 + G4) / 4));
                     if (x == 1)
-                        g.putPixel(x - 1, y, (int) ((G1 + G4 + ip.getPixel(x - 1, y + 1)) / 3));
+                        g.putPixel(x - 1, y,  ((G1 + G4 + ip.getPixel(x - 1, y + 1)) / 3));
                 }
             }
 
@@ -89,15 +89,15 @@ class SmoothDebayerAlgorithm implements DebayerAlgorithm {
                     G3 = ip.getPixel(x + 1, y + 1);
                     G4 = ip.getPixel(x + 1, y - 1);
 
-                    g.putPixel(x, y, (int) G1);
+                    g.putPixel(x, y,  G1);
                     if (x == 0)
-                        g.putPixel(x + 1, y, (int) ((G1 + G2 + G3) / 3));
+                        g.putPixel(x + 1, y,  ((G1 + G2 + G3) / 3));
                     else
-                        g.putPixel(x + 1, y, (int) ((G1 + G2 + G3 + G4) / 4));
+                        g.putPixel(x + 1, y,  ((G1 + G2 + G3 + G4) / 4));
                 }
             }
 
-            g.putPixel(0, 0, (int) ((ip.getPixel(0, 1) + ip.getPixel(1, 0)) / 2));
+            g.putPixel(0, 0,  ((ip.getPixel(0, 1) + ip.getPixel(1, 0)) / 2));
 
             for (int y = 0; y < height; y += 2) {
                 for (int x = 0; x < width; x += 2) {
@@ -121,10 +121,10 @@ class SmoothDebayerAlgorithm implements DebayerAlgorithm {
                     if (G4 == 0)
                         G4 = 1;
 
-                    b.putPixel(x, y, (int) (B1));
-                    b.putPixel(x + 1, y, (int) ((G5 / 2 * ((B1 / G1) + (B2 / G2)))));
-                    b.putPixel(x, y + 1, (int) ((G6 / 2 * ((B1 / G1) + (B3 / G3)))));
-                    b.putPixel(x + 1, y + 1, (int) ((G9 / 4 * ((B1 / G1) + (B3 / G3) + (B2 / G2) + (B4 / G4)))));
+                    b.putPixel(x, y,  (B1));
+                    b.putPixel(x + 1, y,  ((G5 / 2 * ((B1 / G1) + (B2 / G2)))));
+                    b.putPixel(x, y + 1,  ((G6 / 2 * ((B1 / G1) + (B3 / G3)))));
+                    b.putPixel(x + 1, y + 1,  ((G9 / 4 * ((B1 / G1) + (B3 / G3) + (B2 / G2) + (B4 / G4)))));
 
                 }
             }
@@ -151,10 +151,10 @@ class SmoothDebayerAlgorithm implements DebayerAlgorithm {
                     if (G4 == 0)
                         G4 = 1;
 
-                    r.putPixel(x, y, (int) (R1));
-                    r.putPixel(x + 1, y, (int) ((G5 / 2 * ((R1 / G1) + (R2 / G2)))));
-                    r.putPixel(x, y + 1, (int) ((G6 / 2 * ((R1 / G1) + (R3 / G3)))));
-                    r.putPixel(x + 1, y + 1, (int) ((G9 / 4 * ((R1 / G1) + (R3 / G3) + (R2 / G2) + (R4 / G4)))));
+                    r.putPixel(x, y,  (R1));
+                    r.putPixel(x + 1, y,  ((G5 / 2 * ((R1 / G1) + (R2 / G2)))));
+                    r.putPixel(x, y + 1,  ((G6 / 2 * ((R1 / G1) + (R3 / G3)))));
+                    r.putPixel(x + 1, y + 1,  ((G9 / 4 * ((R1 / G1) + (R3 / G3) + (R2 / G2) + (R4 / G4)))));
                 }
             }
 
@@ -178,13 +178,13 @@ class SmoothDebayerAlgorithm implements DebayerAlgorithm {
                     G3 = ip.getPixel(x + 1, y + 1);
                     G4 = ip.getPixel(x + 1, y - 1);
 
-                    g.putPixel(x, y, (int) G1);
+                    g.putPixel(x, y,  G1);
                     if (y == 0)
-                        g.putPixel(x + 1, y, (int) ((G1 + G2 + G3) / 3));
+                        g.putPixel(x + 1, y,  ((G1 + G2 + G3) / 3));
                     else
-                        g.putPixel(x + 1, y, (int) ((G1 + G2 + G3 + G4) / 4));
+                        g.putPixel(x + 1, y,  ((G1 + G2 + G3 + G4) / 4));
                     if (x == 1)
-                        g.putPixel(x - 1, y, (int) ((G1 + G4 + ip.getPixel(x - 1, y + 1)) / 3));
+                        g.putPixel(x - 1, y,  ((G1 + G4 + ip.getPixel(x - 1, y + 1)) / 3));
                 }
             }
 
@@ -195,15 +195,15 @@ class SmoothDebayerAlgorithm implements DebayerAlgorithm {
                     G3 = ip.getPixel(x + 1, y + 1);
                     G4 = ip.getPixel(x + 1, y - 1);
 
-                    g.putPixel(x, y, (int) G1);
+                    g.putPixel(x, y,  G1);
                     if (x == 0)
-                        g.putPixel(x + 1, y, (int) ((G1 + G2 + G3) / 3));
+                        g.putPixel(x + 1, y,  ((G1 + G2 + G3) / 3));
                     else
-                        g.putPixel(x + 1, y, (int) ((G1 + G2 + G3 + G4) / 4));
+                        g.putPixel(x + 1, y,  ((G1 + G2 + G3 + G4) / 4));
                 }
             }
 
-            g.putPixel(0, 0, (int) ((ip.getPixel(0, 1) + ip.getPixel(1, 0)) / 2));
+            g.putPixel(0, 0,  ((ip.getPixel(0, 1) + ip.getPixel(1, 0)) / 2));
 
             for (int y = 1; y < height; y += 2) {
                 for (int x = 0; x < width; x += 2) {
@@ -227,10 +227,10 @@ class SmoothDebayerAlgorithm implements DebayerAlgorithm {
                     if (G4 == 0)
                         G4 = 1;
 
-                    b.putPixel(x, y, (int) (B1));
-                    b.putPixel(x + 1, y, (int) ((G5 / 2 * ((B1 / G1) + (B2 / G2)))));
-                    b.putPixel(x, y + 1, (int) ((G6 / 2 * ((B1 / G1) + (B3 / G3)))));
-                    b.putPixel(x + 1, y + 1, (int) ((G9 / 4 * ((B1 / G1) + (B3 / G3) + (B2 / G2) + (B4 / G4)))));
+                    b.putPixel(x, y,  (B1));
+                    b.putPixel(x + 1, y,  ((G5 / 2 * ((B1 / G1) + (B2 / G2)))));
+                    b.putPixel(x, y + 1,  ((G6 / 2 * ((B1 / G1) + (B3 / G3)))));
+                    b.putPixel(x + 1, y + 1,  ((G9 / 4 * ((B1 / G1) + (B3 / G3) + (B2 / G2) + (B4 / G4)))));
 
                 }
             }
@@ -257,10 +257,10 @@ class SmoothDebayerAlgorithm implements DebayerAlgorithm {
                     if (G4 == 0)
                         G4 = 1;
 
-                    r.putPixel(x, y, (int) (R1));
-                    r.putPixel(x + 1, y, (int) ((G5 / 2 * ((R1 / G1) + (R2 / G2)))));
-                    r.putPixel(x, y + 1, (int) ((G6 / 2 * ((R1 / G1) + (R3 / G3)))));
-                    r.putPixel(x + 1, y + 1, (int) ((G9 / 4 * ((R1 / G1) + (R3 / G3) + (R2 / G2) + (R4 / G4)))));
+                    r.putPixel(x, y,  (R1));
+                    r.putPixel(x + 1, y,  ((G5 / 2 * ((R1 / G1) + (R2 / G2)))));
+                    r.putPixel(x, y + 1,  ((G6 / 2 * ((R1 / G1) + (R3 / G3)))));
+                    r.putPixel(x + 1, y + 1,  ((G9 / 4 * ((R1 / G1) + (R3 / G3) + (R2 / G2) + (R4 / G4)))));
                 }
             }
